requirements.txt
tensorflow-cpu==2.21.0
fastapi==0.135.3
hypercorn==0.18.0
librosa==0.11.0
numpy==2.4.4
python-multipart==0.0.26
opencv-python-headless==4.13.0.92
scikit-learn==1.8.0
typing-extensions==4.15.0
typing-inspection==0.4.2
pydantic==2.13.0
pydantic-core==2.46.0
starlette==1.0.0
anyio==4.13.0
keras==3.14.0
scipy==1.17.1
soundfile==0.13.1
numba==0.65.0
